ref: fix remote origin management
When a flatpakref file does not provide a SuggestRemoteName key,
create an origin from the Name field.

This change implements the algorithm found in flatpak's
create_origin_remote_config() function.

Remove the option of setting an origin label in packageOption.
This option was poorly documented, and its default behavior
- labeling as "flathub" if no origin was set -
would break system activation.