scheduled update won't trigger on activation.
* Refactors the update logic so that when a user enables
`update.auto`, updates will no longer be triggered on activation.

To achieve this, a distinct systemd unit (`flatpak-managed-install-timer`)
will be enabled when auto updates are turned on.

This unit will be executed exclusively by a timer,
and will run a variant of the installer script using the `--or-update`
flatpak flag.

This approach paves the way for further specialization of
onActivation and scheduled execution profiles.

* Refactors the installer api and script generation
into dedicated modules, and introduces unit tests for the
code paths that have been modified.